+++ src/main/java/edu/illinois/library/cantaloupe/config/Key.java
@@ -162,6 +162,8 @@ public enum Key {
     PROCESSOR_BACKGROUND_COLOR("processor.background_color"),
     PROCESSOR_DOWNSCALE_FILTER("processor.downscale_filter"),
     PROCESSOR_DPI("processor.dpi"),
+    PROCESSOR_PDF_SCRATCH_FILE_ENABLED("processor.pdf.scratch_file_enabled"),
+    PROCESSOR_PDF_MAX_MEMORY_BYTES("processor.pdf.max_memory_bytes"),
     PROCESSOR_FALLBACK("processor.ManualSelectionStrategy.fallback"),
     PROCESSOR_FALLBACK_RETRIEVAL_STRATEGY("processor.fallback_retrieval_strategy"),
     PROCESSOR_JPG_PROGRESSIVE("processor.jpg.progressive"),