+++ src/test/java/edu/illinois/library/cantaloupe/processor/PdfBoxProcessorTest.java
@@ -28,6 +28,9 @@ public class PdfBoxProcessorTest extends AbstractProcessorTest {
         super.setUp();
         Configuration config = Configuration.getInstance();
         config.setProperty(Key.PROCESSOR_DPI, 72);
+        config.setProperty(Key.PROCESSOR_PDF_SCRATCH_FILE_ENABLED, false);
+        config.setProperty(Key.PROCESSOR_PDF_MAX_MEMORY_BYTES, -1);
+        config.setProperty(Key.TEMP_PATHNAME, "/tmp");
         instance = newInstance();
     }